requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
numpy==2.1.1
python-dotenv==1.0.1
requests==2.32.3
scipy==1.14.1
urllib3==2.2.3
